Move random string helpers into a separate file
