+++ spec/helpers/random_helpers.h
@@ -1,0 +1,8 @@
+#ifndef HELPERS_RANDOM_HELPERS_H_
+#define HELPERS_RANDOM_HELPERS_H_
+
+#include <string>
+
+std::string random_words(size_t count);
+
+#endif  // HELPERS_RANDOM_HELPERS_H_
